Fixing hardware timing issues for i2c
I came across the same issue as described here: https://devzone.nordicsemi.com/f/nordic-q-a/48166/i2c-clock-stretching-on-nrf52840
This hack should prevent the first clock after a clock stretch to be too short. We initially started our investigation because of random bit shifts ( >> 1) in the communication with BQ40Z50 which resulted in faulty values. In case of fuel gauges and chargers this can lead to big problems (settings wrong currents, voltages, ...).